Moved Board to /components
